+++ features/steps/git_steps.py
@@ -192,6 +192,21 @@ def step_impl(context, name, ending, filename, gitattr):
         tag("v1")
 
 
+@given('a git-repository "{name}" with a {ending} "{filename}" file')
+def step_impl(context, name, ending, filename):
+    terminator = {"LF": "\n", "CRLF": "\r\n"}[ending]
+    remote_path = os.path.join(context.remotes_dir, name)
+    pathlib.Path(remote_path).mkdir(parents=True, exist_ok=True)
+    with in_directory(remote_path):
+        create_repo()
+        subprocess.check_call(["git", "config", "core.autocrlf", "false"])
+        pathlib.Path(filename).write_bytes(
+            f"Generated file for {name}{terminator}".encode("utf-8")
+        )
+        commit_all("Initial commit")
+        tag("v1")
+
+
 @given("all files in {directory} are committed")
 @when("all files in {directory} are committed")
 def step_impl(_, directory):